controlgraph/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.0
pydantic>=2.7.0
pydantic-settings>=2.2.1
python-jose[cryptography]>=3.3.0
passlib>=1.7.4
email-validator>=2.1.1
requests>=2.32.0
httpx>=0.27.0
pandas>=2.2.0
matplotlib>=3.8.0
psycopg[binary]>=3.1.19
